tools/history-clip-tool/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Video processing
ffmpeg-python==0.2.0

# Transcription
faster-whisper>=1.0.0
torch>=2.1.2  # CPU-only version
torchaudio>=2.1.2

# Optional face detection
opencv-python==4.9.0.80

# Database
sqlalchemy==2.0.25

# Configuration
toml==0.10.2

# Utilities
python-multipart==0.0.6  # File uploads
aiofiles==23.2.1  # Async file operations

# GUI (for standalone executable)
pywebview==4.4.1  # Native window for UI

# Packaging (for building executable)
pyinstaller==6.3.0
